requirements.txt
selenium>=4.16.0
axe-selenium-python>=2.1.6
opencv-python-headless>=4.8.0
transformers[torch]>=4.34.0
torch>=2.0.0
streamlit>=1.28.0
reportlab>=4.0.4
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=7.4.0
datasets>=2.14.0
webdriver-manager>=4.0.0
Pillow>=10.0.0
google-generativeai>=0.3.0
